+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This document contains what changed in each release of Fisher.
 
 *Not yet released.*
 
+* Add support for working behind proxies
 * GH-1: Add support for receiving hooks from **GitLab**
 * Improve unit testing coverage of the project